requirements.txt
fastapi
uvicorn
pydantic
python-multipart
aiohttp
pypdf
langchain
langchain-community
python-dotenv
typing-extensions
google-genai==1.17.0
